ClangComplexityBear: Move to c_languages
This has nothing to do with the code clone detection.
