+++ tests/c_languages/test_files/empty_declarations.c
